fix: add translations to the ErrorBoundary component
