+++ src/components/shared/primitives/__stories__/Text.stories.tsx
@@ -21,7 +21,7 @@ storiesOf('Text', module)
         variant={select(
           'Variant',
           ['small', 'medium', 'regular', 'large'],
-          'large'
+          'large',
         )}
       >
         Large variant
